// src/app/shared/types/api-types/bire-modification-dto.interface.ts
export interface BireModificationDTO {
  modificationType?: string;
  modificationNumber?: string;
  modificationRevisionNumber?: string;
  modificationTitle?: string;
  familyCode?: string;
  structureType?: string;
  modificationDocumentUrl?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}