// src/app/shared/types/api-types/bire-modification-evolution-dto.interface.ts
export interface BireModificationEvolutionDTO {
  modificationType?: string;
  modificationTitle?: string;
  modificationNumber?: string;
  modificationRevisionNumber?: string;
  evolutionNumber?: string;
  evolutionRevisionNumber?: string;
  familyCode?: string;
  structureType?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}